# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API REST para reservas, servicios, combos y administracion de Atrevida Fit.",
        "title": "Atrevida Fit - Agenda API",
        "contact": {},
        "version": "v0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/reservas": {
            "get": {
                "description": "Devuelve reservas desde Google Sheets con filtros. Filtros: local (opcional), semana YYYY-MM-DD (opcional), dia lunes..sabado (opcional), tipo mesa/bicicleta/feriado (opcional), cliente (opcional), reservados true/false (opcional). Response: total_locales (int), filtros (objeto con filtros aplicados), reservas ([]LocalReservas cada uno con: local string, semanas []Semana con titulo y slots []ReservaSlot con hora y map dia-\u003e[]ReservaItem con tipo M/B, cliente, servicio, servicio_solicitado, servicio_confirmado, estado, numero_telefono, notificado, creado_en, actualizado_en).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reservas Sheets"
                ],
                "summary": "Listar reservas desde Google Sheets",
                "parameters": [
                    {
                        "type": "string",
                        "example": "SAN MARTIN",
                        "description": "Nombre del local",
                        "name": "local",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2026-05-25",
                        "description": "Semana a consultar YYYY-MM-DD",
                        "name": "semana",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "lunes",
                        "description": "Dia a consultar",
                        "name": "dia",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "mesa",
                            "bicicleta",
                            "feriado"
                        ],
                        "type": "string",
                        "example": "mesa",
                        "description": "Tipo de reserva",
                        "name": "tipo",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "Maria",
                        "description": "Nombre del cliente",
                        "name": "cliente",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "example": true,
                        "description": "Filtrar solo reservados",
                        "name": "reservados",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.reservaListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error de validacion: tipo invalido",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Crea una reserva en Google Sheets. local: nombre del local (requerido). semana: YYYY-MM-DD (requerido). dia: lunes a sabado (requerido). hora_desde: HH:MM (requerido). hora_hasta: HH:MM (opcional). tipo: M=mesa o B=bicicleta (requerido). cliente: nombre del cliente (requerido). servicio: nombre del servicio (opcional). Si hay conflictos, algunos slots fallan (207).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reservas Sheets"
                ],
                "summary": "Crear reserva en Google Sheets",
                "parameters": [
                    {
                        "description": "Datos de la reserva",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.crearReservaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todos los slots reservados exitosamente",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.slotsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "207": {
                        "description": "Algunos slots no pudieron reservarse",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.slotsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error de validacion: campo requerido faltante o tipo invalido",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Conflicto: ningun slot pudo reservarse",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Modifica una reserva en Google Sheets. Solo se actualizan los campos nuevos enviados. local: nombre del local (requerido). semana: YYYY-MM-DD (requerido). dia: lunes a sabado (requerido). hora: hora actual HH:MM (requerido). tipo: M=mesa o B=bicicleta (requerido). cliente: nombre del cliente (requerido). nuevo_dia: nuevo dia (opcional). nueva_hora_desde: nueva hora inicio (opcional). nueva_hora_hasta: nueva hora fin (opcional). nuevo_tipo: M o B (opcional). nuevo_servicio: nuevo servicio (opcional). Si hay conflictos, algunos slots fallan (207).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reservas Sheets"
                ],
                "summary": "Actualizar reserva en Google Sheets",
                "parameters": [
                    {
                        "description": "Datos para actualizar la reserva",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.actualizarReservaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todos los slots actualizados exitosamente",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.slotsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "207": {
                        "description": "Algunos slots no pudieron actualizarse",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.slotsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error de validacion: campo requerido faltante, tipo invalido, sin cambios para actualizar",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Conflicto: ningun slot pudo actualizarse",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.actualizarReservaRequest": {
            "type": "object",
            "required": [
                "cliente",
                "dia",
                "hora",
                "local",
                "semana",
                "tipo"
            ],
            "properties": {
                "cliente": {
                    "description": "Nombre del cliente",
                    "type": "string",
                    "example": "Maria Lopez"
                },
                "dia": {
                    "description": "Dia actual de la reserva",
                    "type": "string",
                    "example": "lunes"
                },
                "hora": {
                    "description": "Hora actual de la reserva",
                    "type": "string",
                    "example": "09:00"
                },
                "local": {
                    "description": "Nombre del local",
                    "type": "string",
                    "example": "SAN MARTIN"
                },
                "nueva_hora_desde": {
                    "description": "Nueva hora de inicio (opcional)",
                    "type": "string",
                    "example": "11:00"
                },
                "nueva_hora_hasta": {
                    "description": "Nueva hora de fin (opcional)",
                    "type": "string",
                    "example": "12:00"
                },
                "nuevo_dia": {
                    "description": "Nuevo dia (opcional)",
                    "type": "string",
                    "example": "martes"
                },
                "nuevo_servicio": {
                    "description": "Nuevo nombre del servicio (opcional)",
                    "type": "string",
                    "example": "Evaluacion corporal"
                },
                "nuevo_tipo": {
                    "description": "Nuevo tipo de espacio: M o B (opcional)",
                    "type": "string",
                    "example": "B"
                },
                "semana": {
                    "description": "Semana en formato YYYY-MM-DD",
                    "type": "string",
                    "example": "2026-05-25"
                },
                "tipo": {
                    "description": "Tipo actual de espacio: M (mesa) o B (bicicleta)",
                    "type": "string",
                    "example": "M"
                }
            }
        },
...
        "handlers.crearReservaRequest": {
            "type": "object",
            "required": [
                "cliente",
                "dia",
                "hora_desde",
                "local",
                "semana",
                "tipo"
            ],
            "properties": {
                "cliente": {
                    "description": "Nombre del cliente",
                    "type": "string",
                    "example": "Maria Lopez"
                },
                "dia": {
                    "description": "Dia de la semana (lunes, martes, miercoles, jueves, viernes, sabado)",
                    "type": "string",
                    "example": "lunes"
                },
                "hora_desde": {
                    "description": "Hora de inicio (HH:MM)",
                    "type": "string",
                    "example": "09:00"
                },
                "hora_hasta": {
                    "description": "Hora de fin (HH:MM), opcional",
                    "type": "string",
                    "example": "10:00"
                },
                "local": {
                    "description": "Nombre del local",
                    "type": "string",
                    "example": "SAN MARTIN"
                },
                "semana": {
                    "description": "Semana en formato YYYY-MM-DD (primer dia de la semana, lunes)",
                    "type": "string",
                    "example": "2026-05-25"
                },
                "servicio": {
                    "description": "Nombre del servicio (ej: \"Depilacion piernas completas\"). Texto libre.",
                    "type": "string",
                    "example": "Depilacion piernas completas"
                },
                "tipo": {
                    "description": "Tipo de espacio: M (mesa) o B (bicicleta)",
                    "type": "string",
                    "example": "M"
                }
            }
        },
...
        "handlers.reservaFiltrosResponse": {
            "type": "object",
            "properties": {
                "cliente": {
                    "type": "string",
                    "example": "Maria"
                },
                "dia": {
                    "type": "string",
                    "example": "lunes"
                },
                "local": {
                    "type": "string",
                    "example": "SAN MARTIN"
                },
                "reservados": {
                    "type": "boolean",
                    "example": true
                },
                "semana": {
                    "type": "string",
                    "example": "2026-05-25"
                },
                "tipo": {
                    "type": "string",
                    "example": "mesa"
                }
            }
        },
...
        "handlers.reservaListResponse": {
            "type": "object",
            "properties": {
                "filtros": {
                    "$ref": "#/definitions/handlers.reservaFiltrosResponse"
                },
                "reservas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.LocalReservas"
                    }
                },
                "total_locales": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
...
        "handlers.slotsResponse": {
            "type": "object",
            "properties": {
                "slots_error": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "11:00",
                        "12:00"
                    ]
                },
                "slots_ok": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "09:00",
                        "10:00"
                    ]
                }
            }
        },
...
        "models.LocalReservas": {
            "type": "object",
            "properties": {
                "local": {
                    "type": "string",
                    "example": "SAN MARTIN"
                },
                "semanas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Semana"
                    }
                }
            }
        },
        "models.ReservaItem": {
            "type": "object",
            "properties": {
                "actualizado_en": {
                    "type": "string",
                    "example": "2026-05-23T16:04:05Z"
                },
                "cliente": {
                    "type": "string",
                    "example": "Maria Lopez"
                },
                "creado_en": {
                    "type": "string",
                    "example": "2026-05-23T15:04:05Z"
                },
                "estado": {
                    "type": "string",
                    "example": "AGENDADO"
                },
                "notificado": {
                    "type": "boolean",
                    "example": false
                },
                "numero_telefono": {
                    "type": "string",
                    "example": "+59170011223"
                },
                "servicio": {
                    "type": "string",
                    "example": "Depilacion Laser"
                },
                "servicio_confirmado": {
                    "type": "string",
                    "example": "Depilacion Laser Piernas"
                },
                "servicio_solicitado": {
                    "type": "string",
                    "example": "Piernas completas"
                },
                "tipo": {
                    "type": "string",
                    "example": "M"
                }
            }
        },
        "models.ReservaSlot": {
            "type": "object",
            "properties": {
                "dias": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/models.ReservaItem"
                        }
                    }
                },
                "hora": {
                    "type": "string",
                    "example": "09:00"
                }
            }
        },
        "models.Semana": {
            "type": "object",
            "properties": {
                "reservas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ReservaSlot"
                    }
                },
                "titulo": {
                    "type": "string",
                    "example": "Semana 22"
                }
            }
        },
...
        "utils.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "error": {
                    "type": "boolean",
                    "example": false
                },
                "message": {
                    "type": "string",
                    "example": "operacion realizada correctamente"
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                }
            }
        }
    }
}